refactor : AccommodationRequest 이동으로 인한 import 수정
